rename history to blockstore
