+++ blockstore/src/block_stores/mod.rs
@@ -1,3 +1,3 @@
-pub mod faithful;
+pub mod faithful_history;
 pub mod multiple_strategy_block_store;
 pub mod postgres;
